% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.R
\name{abs_angle}
\alias{abs_angle}
\title{Calculate Absolute Angles}
\usage{
abs_angle(DT, coords, datetime, by, returnIntermediate = FALSE)
}
\arguments{
\item{DT}{data.table}

\item{coords}{character vector; coordinate columns}

\item{datetime}{character; datetime column}

\item{by}{character vector; columns identifying within which groups (e.g. each individual and year) to calculate angles}

\item{returnIntermediate}{boolean; return intermediate columns}
}
\value{
input DT appended with absolute angle column
}
\description{
Calculate Absolute Angles
}
