% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_abs_angle.R
\name{calc_abs_angle}
\alias{calc_abs_angle}
\title{Absolute angle}
\usage{
calc_abs_angle(DT, coords, datetime, id, yr, allCW = FALSE,
  returnIntermediate = FALSE)
}
\arguments{
\item{DT}{data.table}

\item{coords}{character vector; coordinate columns}

\item{datetime}{character; datetime column}

\item{id}{character; id column}

\item{yr}{character; year column}

\item{allCW}{boolean; all clockwise}

\item{returnIntermediate}{boolean; return intermediate columns}
}
\value{
input DT appended with absolute angle column
}
\description{
Absolute angle
}
