% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_di.R
\name{calc_di}
\alias{calc_di}
\title{Calculate DI}
\usage{
calc_di(DT, suffix, angle, dist, alpha = 1, radians)
}
\arguments{
\item{DT}{}

\item{suffix}{character to append directly to angle column name, indicating the angle of the neighbour.}

\item{dist}{}

\item{alpha}{default is 1.}

\item{radians}{TRUE/FALSE if angles provided are in radians (TRUE) or in degrees (FALSE)}
}
\description{
Dynamic interaction as described by Long and Nelson (2013).
}
\references{
Long, Jed A., and Trisalyn A. Nelson. "Measuring dynamic interaction in movement data." Transactions in GIS 17.1 (2013): 62-77.
}
