% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rel_angle.R
\name{calc_rel_angle}
\alias{calc_rel_angle}
\title{Calculate Relative Angles}
\usage{
calc_rel_angle(DT, coords = NULL, datetime = NULL, by = NULL)
}
\arguments{
\item{DT}{data.table}

\item{coords}{character vector; coordinate columns}

\item{datetime}{character; datetime column}

\item{by}{character vector; columns identifying within which groups (e.g. each individual and year) to calculate angles}
}
\description{
Calculate Relative Angles
}
