% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_di.R
\name{diff_azimuth}
\alias{diff_azimuth}
\title{Difference in azimuth}
\usage{
diff_azimuth(DT, angle1, angle2)
}
\arguments{
\item{DT}{data.table}

\item{angle1}{character of column name with azimuth angle for left individual}

\item{angle2}{character of column name with  azimuth angle for right individual}
}
\description{
Difference in azimuth or absolute angle, calculated with \code{\link{abs_angle}}.
}
\seealso{
abs_angle
}
