% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_di.R
\name{diff_dist}
\alias{diff_dist}
\title{Difference in distance}
\usage{
diff_dist(DT, suffix, dist, alpha = 1)
}
\arguments{
\item{DT}{data.table}

\item{suffix}{character to append directly to angle column name, indicating the angle of the neighbour.}

\item{dist}{}

\item{alpha}{default is 1.}
}
\value{

}
\description{
Difference in distance
}
\references{
Long, Jed A., and Trisalyn A. Nelson. "Measuring dynamic interaction in movement data." Transactions in GIS 17.1 (2013): 62-77.
}
