% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyad_dist.R
\name{dyad_dist}
\alias{dyad_dist}
\title{Calculate dyad distance}
\usage{
dyad_dist(DT, coords, suffix = NULL, returnIntermediate = FALSE)
}
\arguments{
\item{returnIntermediate}{}
}
\description{
For each dyad, calculate the euclidean distance between locations.
}
