% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grid.R
\name{generate_grid}
\alias{generate_grid}
\title{Generate grid}
\usage{
generate_grid(pol, spacing, crs)
}
\arguments{
\item{pol}{mcp or other SpatialPolygons object within which the grid will be generated.}

\item{spacing}{distance between points in grid. units of the projection.}

\item{crs}{proj4string of coordinate reference system for pol.}
}
\value{

}
\description{
Generate a regular grid within a polygon object.
}
