% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumbQuadTreeNeighbours.R
\name{NumbQuadTreeNeighbours}
\alias{NumbQuadTreeNeighbours}
\title{Title}
\usage{
NumbQuadTreeNeighbours(DT, coords, numbNeighbours, idCol)
}
\arguments{
\item{DT}{}

\item{coords}{}

\item{numbNeighbours}{}

\item{idCol}{}
}
\description{
Title
}
