% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_date.R
\name{prep_date}
\alias{prep_date}
\title{Date prep}
\usage{
prep_date(DT, dateCol, timeCol, dateFormat = NULL, timeFormat = NULL)
}
\arguments{
\item{DT}{data.table}

\item{dateCol}{}

\item{timeCol}{}

\item{dateFormat}{}

\item{timeFormat}{}
}
\description{
Date prep
}
