% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rel_angle.R
\name{calc_rel_angle}
\alias{calc_rel_angle}
\title{Relative Angles}
\usage{
calc_rel_angle(DT, coords = NULL, datetime = NULL, id = NULL,
  yr = NULL, allCW = FALSE, returnIntermediate = FALSE)
}
\arguments{
\item{DT}{data.table}

\item{coords}{character vector; coordinate columns}

\item{datetime}{character; datetime column}

\item{id}{character; id column}

\item{yr}{character; year column}

\item{allCW}{boolean; all clockwise}

\item{returnIntermediate}{boolean; return intermediate columns}
}
\description{
Relative Angles
}
