% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rel_angle.R
\name{calc_rel_angle}
\alias{calc_rel_angle}
\title{Relative Angles}
\usage{
calc_rel_angle(DT, abs = NULL, coord = NULL, datetime = NULL,
  id = NULL, allCW = FALSE, returnIntermediate = FALSE)
}
\arguments{
\item{DT}{}

\item{abs}{}

\item{coord}{}

\item{datetime}{}

\item{id}{}

\item{allCW}{}

\item{returnIntermediate}{}
}
\description{
Relative Angles
}
