% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_di.R
\name{diff_azimuth}
\alias{diff_azimuth}
\title{Difference in azimuth}
\usage{
diff_azimuth(DT, suffix, angle)
}
\arguments{
\item{DT}{data.table}

\item{suffix}{character to append directly to angle column name, indicating the angle of the neighbour.}

\item{angles}{character of column name with  azimuth angle.}
}
\value{

}
\description{
Difference in azimuth or absolute angle, calculated with \code{\link{abs_angle}}.
}
\references{
Long, Jed A., and Trisalyn A. Nelson. "Measuring dynamic interaction in movement data." Transactions in GIS 17.1 (2013): 62-77.
}
