% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyad_dist.R
\name{dyad_dist}
\alias{dyad_dist}
\title{Calculate dyad distance}
\usage{
dyad_dist(DT, coords, suffix = NULL, returnIntermediate = FALSE)
}
\arguments{
\item{DT}{data.table}

\item{coords}{character vector; coordinate columns}

\item{suffix}{in column names associated with neighbours (e.g.: '.nn')}

\item{returnIntermediate}{boolean; return intermediate columns}
}
\value{

}
\description{
For each dyad, calculate the euclidean distance between locations.
}
