% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_date.R
\name{prep_date}
\alias{prep_date}
\title{Date prep}
\usage{
prep_date(DT, dateCol, timeCol, tz, dateFormat = NULL, timeFormat = NULL)
}
\arguments{
\item{DT}{data.table}

\item{dateCol}{character of date column name}

\item{timeCol}{character of time column name}

\item{dateFormat}{character of date format as described in `strptime`}

\item{timeFormat}{character of time format as described in `strptime`}
}
\value{

}
\description{
Date prep
}
